Use Int vs Float for money, always !

+++ app/models/PaymentMethod.scala
@@ -15,7 +15,7 @@ import scala.util.{Try, Failure, Success}
 import com.stripe.model.{Card => StripeCard}
 
 abstract class PaymentMethod {
-  def authenticate(amount: Float)(implicit ec: ExecutionContext): Future[String Or List[ErrorMessage]]
+  def authenticate(amount: Int)(implicit ec: ExecutionContext): Future[String Or List[ErrorMessage]]
 }
 
 object PaymentMethods {
@@ -43,12 +43,12 @@ case object FailedDebit extends GiftCardPaymentStatus
 
 // TODO: Figure out how to have the 'status' field on the payment and not the payment method.
 case class CreditCard(id: Int, orderId: Int, cardholderName: String, cardNumber: String, cvv: Int, status: CreditCardPaymentStatus, expiration: String, address: Address) extends PaymentMethod {
-  def authenticate(amount: Float)(implicit ec: ExecutionContext): Future[String Or List[ErrorMessage]] = {
+  def authenticate(amount: Int)(implicit ec: ExecutionContext): Future[String Or List[ErrorMessage]] = {
     Future.successful(Good("authenticated"))
   }
 }
 case class GiftCard(id: Int, orderId: Int, status: GiftCardPaymentStatus, code: String) extends PaymentMethod {
-  def authenticate(amount: Float)(implicit ec: ExecutionContext): Future[String Or List[ErrorMessage]] = {
+  def authenticate(amount: Int)(implicit ec: ExecutionContext): Future[String Or List[ErrorMessage]] = {
     Future.successful(Good("authenticated"))
   }
 }
